fix: exports for extracted getAddWorkflowConfig
