+++ packages/common/src/associations/getAvailableToRequestEntitiesQuery.ts
@@ -7,7 +7,7 @@ import { getProp } from "../objects/get-prop";
 import { getTypesFromEntityType } from "../core/getTypesFromEntityType";
 import { getIdsFromKeywords } from "./internal/getIdsFromKeywords";
 import { getTypeByNotIdsQuery } from "./internal/getTypeByNotIdsQuery";
-import { negateGroupPredicates } from "../search/_internal/negateGroupPredicates";
+import { negateGroupPredicates } from "../search/negateGroupPredicates";
 import { getTypeByIdsQuery } from "./internal/getTypeByIdsQuery";
 import { combineQueries } from "../search/_internal/combineQueries";
