+++ packages/common/src/search/_internal/index.ts
@@ -1,10 +1,8 @@
-export * from "./portalSearchItems";
+export * from "./getWorkflowForType";
+export * from "./hubSearchChannels";
+export * from "./hubSearchEventAttendees";
+export * from "./hubSearchEvents";
 export * from "./hubSearchItems";
 export * from "./portalSearchGroups";
+export * from "./portalSearchItems";
 export * from "./portalSearchUsers";
-export * from "./hubSearchChannels";
-export * from "./hubSearchEvents";
-export * from "./hubSearchEventAttendees";
-export * from "./getWorkflowForType";
-export * from "./getCatalogGroups";
-export * from "./getUserGroupsByMembership";
