+++ packages/common/src/search/index.ts
@@ -1,8 +1,11 @@
-export * from "./utils";
-export * from "./types";
-export * from "./upgradeCatalogSchema";
-export * from "./serializeQueryForPortal";
-export * from "./hubSearch";
 export * from "./Catalog";
 export * from "./Collection";
+export * from "./getUserGroupsByMembership";
+export * from "./getUserGroupsFromQuery";
+export * from "./hubSearch";
+export * from "./negateGroupPredicates";
+export * from "./serializeQueryForPortal";
+export * from "./types";
+export * from "./upgradeCatalogSchema";
+export * from "./utils";
 export * from "./wellKnownCatalog";
