+++ packages/common/src/search/types/AddContentWorkflowTypes.ts
@@ -2,12 +2,14 @@ import { IGroupsByMembership } from "./IGroupsByMembership";
 import { EntityType, IQuery } from "./IHubCatalog";
 
 /**
+ * DEPRECATED - moved into hub-components
  * `AddContentWorkflow` is a string literal type that
  * defines the possible "Add Content" workflows
  */
 export type AddContentWorkflow = "create" | "upload" | "existing";
 
 /**
+ * DEPRECATED - moved into hub-components
  * `IAddContentWorkflowConfig` is an interface that defines the
  * configuration for the "Add Content" workflows
  */
@@ -24,12 +26,13 @@ export interface IAddContentWorkflowConfig {
     | "unsupported-target-entity";
   // FUTURE we can add the checks
 }
-
+/**
+ * DEPRECATED - moved into hub-components
+ */
 export type AddContentWorkflowConfig =
   | IAddContentCreateWorkflowConfig
   | IAddContentExistingWorkflowConfig
   | IAddContentUploadWorkflowConfig;
-
 interface IAddContentWorkflowBaseConfig {
   targetEntity: EntityType;
   workflow: AddContentWorkflow;
@@ -37,17 +40,26 @@ interface IAddContentWorkflowBaseConfig {
   // If groups are not passed in then the user can choose any group they are a member of
   groups?: IGroupsByMembership;
 }
+/**
+ * DEPRECATED - moved into hub-components
+ */
 export interface IAddContentCreateWorkflowConfig
   extends IAddContentWorkflowBaseConfig {
   workflow: "create";
 }
 
+/**
+ * DEPRECATED - moved into hub-components
+ */
 export interface IAddContentExistingWorkflowConfig
   extends IAddContentWorkflowBaseConfig {
   workflow: "existing";
   query: IQuery;
 }
 
+/**
+ * DEPRECATED - moved into hub-components
+ */
 export interface IAddContentUploadWorkflowConfig
   extends IAddContentWorkflowBaseConfig {
   workflow: "upload";
